+++ src/editor-mathfield/render.ts
@@ -269,12 +269,15 @@ export function renderSelection(
 
   // Logic to accommodate mathfield hosted in an isotropically scale-transformed element.
   // Without this, the selection indicator will not be in the right place.
+
   // 1. Inquire how big the mathfield thinks it is
-  const supposedWidth = parseFloat(getComputedStyle(field).width);
+  const offsetWidth = field.offsetWidth;
+
   // 2. Get the actual screen width of the box
   const actualWidth = field.getBoundingClientRect().width;
+
   // 3. Divide the two to get the scale factor
-  let scaleFactor = actualWidth / supposedWidth;
+  let scaleFactor = Math.floor(actualWidth) / offsetWidth;
   scaleFactor = isNaN(scaleFactor) ? 1 : scaleFactor;
 
   if (model.selectionIsCollapsed) {